requirements.txt
# Core Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
sqlite3

# AI/ML Libraries
openai==1.3.0
transformers==4.35.0
torch==2.1.0
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11
python-magic==0.4.27
pandas==2.1.3

# Web Scraping & Text Processing
beautifulsoup4==4.12.2
requests==2.31.0
textstat==0.7.3

# Authentication & Security
Werkzeug==2.3.7
bcrypt==4.0.1
PyJWT==2.8.0

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1

# Development
pytest==7.4.3
black==23.9.1
flake8==6.1.0
